Move python_path to commands.
