+++ hatch/cli.py
@@ -7,7 +7,7 @@ import click
 
 from hatch.commands import (
     conda, python, config, new, init, install, uninstall, grow, update, test,
-    build, clean, release
+    build, clean, release, python_path
 )
 from hatch.commands.utils import (
     CONTEXT_SETTINGS, UNKNOWN_OPTIONS, echo_failure, echo_info, echo_success,
@@ -56,70 +56,7 @@ hatch.add_command(uninstall)
 hatch.add_command(update)
 hatch.add_command(clean)
 hatch.add_command(release)
-
-
-def list_pypaths(ctx, param, value):
-    if not value or ctx.resilient_parsing:
-        return
-
-    try:
-        settings = load_settings()
-    except FileNotFoundError:
-        echo_failure('Unable to locate config file. Try `hatch config --restore`.')
-        sys.exit(1)
-
-    pypaths = settings.get('pypaths', {})
-    if pypaths:
-        for p in pypaths:
-            echo_success('{} -> '.format(p), nl=False)
-            echo_info('{}'.format(pypaths[p]))
-    else:
-        echo_failure('There are no saved Python paths. Add '
-                     'one via `hatch pypath NAME PATH`.')
-
-    ctx.exit()
-
-
-@hatch.command('pypath', context_settings=CONTEXT_SETTINGS,
-               short_help='Names a Python path or shows available ones')
-@click.argument('name')
-@click.argument('path')
-@click.option('-l', '--list', 'show', is_flag=True, is_eager=True, callback=list_pypaths,
-              help='Shows available Python paths.')
-def python_path(name, path, show):
-    """Names an absolute path to a Python executable. You can also modify
-    these in the config file entry `pypaths`.
-
-    Hatch can then use these paths by name when creating virtual envs, building
-    packages, etc.
-
-    \b
-    $ hatch pypath -l
-    There are no saved Python paths. Add one via `hatch pypath NAME PATH`.
-    $ hatch pypath py2 /usr/bin/python
-    Successfully saved Python `py2` located at `/usr/bin/python`.
-    $ hatch pypath py3 /usr/bin/python3
-    Successfully saved Python `py3` located at `/usr/bin/python3`.
-    $ hatch pypath -l
-    py2 -> /usr/bin/python
-    py3 -> /usr/bin/python3
-    """
-    try:
-        settings = load_settings()
-    except FileNotFoundError:
-        echo_failure('Unable to locate config file. Try `hatch config --restore`.')
-        sys.exit(1)
-
-    if 'pypaths' not in settings:
-        updated_settings = copy_default_settings()
-        updated_settings.update(settings)
-        settings = updated_settings
-        echo_success('Settings were successfully updated to include `pypaths` entry.')
-
-    settings['pypaths'][name] = path
-    save_settings(settings)
-    echo_success('Successfully saved Python `{}` located at `{}`.'.format(name, path))
-
+hatch.add_command(python_path)
 
 def list_envs(ctx, param, value):
     if not value or ctx.resilient_parsing:
